+++ cli/src/commands/extensions/extension.rs
@@ -15,7 +15,6 @@ use serde::{Deserialize, Serialize};
 use walkdir::WalkDir;
 
 use crate::api::PhylumApi;
-pub(crate) use crate::commands::extensions::api::ExtensionState;
 use crate::commands::extensions::permissions::Permissions;
 use crate::commands::{CommandResult, ExitCode};
 use crate::{deno, dirs};
